requirements.txt
librosa
opencv-python
numpy
moviepy
matplotlib
Pillow
mediapipe